+++ android/src/main/java/com/bitmovin/player/reactnative/extensions/Bitmap.kt
@@ -8,4 +8,5 @@ fun Bitmap.toBase64DataUri(): String {
     val byteArrayOutputStream = ByteArrayOutputStream()
     this.compress(Bitmap.CompressFormat.PNG, 100, byteArrayOutputStream)
     val byteArray = byteArrayOutputStream.toByteArray()
-    return "data:image/png;base64," + Base64.encodeToString(byteArray, Base64.NO_WRAP)}
+    return "data:image/png;base64," + Base64.encodeToString(byteArray, Base64.NO_WRAP)
+}