+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-- `CueEnterEvent.imageUri` and `CueExitEvent.imageUri` to expose the Base64 encoded image data URI of the cue when available
+- `CueEnterEvent.image` and `CueExitEvent.image` to expose the Base64 encoded image data URI of the cue when available
 
 ## [0.38.0] - 2025-02-28